requirements-runtime.txt
onnxruntime
numpy==1.26.4
pillow==10.3.0
fastapi
uvicorn[standard]  # [standard] для підтримки websockets та іншого, якщо знадобиться
python-multipart # Для обробки завантаження файлів у FastAPI